# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS API Documentation",
        "description": "HRMS Backend API documentation with OpenAPI/Swagger",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/user/profile-picture": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile picture",
                "description": "Update user profile picture.",
                "operationId": "220a790c95b99067df27724868ba24fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_picture"
                                ],
                                "properties": {
                                    "profile_picture": {
                                        "description": "Profile picture file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile picture updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile picture updated successfully"
                                        },
                                        "profile_picture": {
                                            "type": "string",
                                            "example": "profile_pictures/image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/username": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user name",
                "description": "Update user name.",
                "operationId": "3e27ac34580dd53b9d5582c2d5774c7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Username updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Username updated successfully"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/email": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user email",
                "description": "Update user email.",
                "operationId": "2fe0123db0eff590c3b180dc268eba53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email updated successfully"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user password",
                "description": "Update user password.",
                "operationId": "37ec0ac34a5eb75f5988a70316e39f7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "current123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new123"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Current password is incorrect"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenticated user details",
                "description": "Returns the authenticated user's basic details along with their roles and permissions.",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "id",
                                        "name",
                                        "email",
                                        "roles",
                                        "permissions"
                                    ],
                                    "properties": {
                                        "id": {
                                            "description": "Unique identifier for the user",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Full name of the user",
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "description": "Email address of the user",
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        },
                                        "last_login_at": {
                                            "description": "The date and time when the user last logged in",
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "roles": {
                                            "description": "List of roles assigned to the user",
                                            "type": "array",
                                            "items": {
                                                "description": "A role name",
                                                "type": "string",
                                                "example": "Admin"
                                            }
                                        },
                                        "permissions": {
                                            "description": "List of permissions granted to the user",
                                            "type": "array",
                                            "items": {
                                                "description": "A permission identifier",
                                                "type": "string",
                                                "example": "user.read"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - The request does not have valid authentication credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}